requirements.txt
pm4py==2.2.32
networkx
pandas
numpy
openpyxl
matplotlib
seaborn
tensorflow-gpu